requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
confluent-kafka==1.9.2
Faker==14.2.0
idna==3.3
oauthlib==3.2.1
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.10.1
typing_extensions==4.3.0
urllib3==1.26.12
